11.3.3.102 Official Release - Details can be found in ReadMe and NEWS
