+++ mstrio/__init__.py
@@ -1,6 +1,6 @@
 __title__ = "mstrio-py"
-__version__ = "11.3.3.101"  # NOSONAR
-__dev_version__ = 1019
+__version__ = "11.3.3.102"  # NOSONAR
+__dev_version__ = 1027
 __license__ = "Apache License 2.0"
 __description__ = "Python interface for the MicroStrategy REST API"
 __author__ = "MicroStrategy"
