+++ mstrio/distribution_services/subscription/__init__.py
@@ -1,6 +1,8 @@
 # flake8: noqa
+from .base_subscription import Subscription
+from .cache_update_subscription import CacheUpdateSubscription
 from .content import Content
 from .delivery import (CacheType, ClientType, Delivery, Orientation, SendContentAs,
                        ShortcutCacheFormat, ZipSettings)
-from .subscription import EmailSubscription, list_subscriptions, Subscription
-from .subscription_manager import SubscriptionManager
+from .email_subscription import EmailSubscription
+from .subscription_manager import list_subscriptions, SubscriptionManager
